+++ resources/js/components/GlobalNav.jsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import {Link} from 'react-router-dom';
+import React from "react";
+import { Link } from "react-router-dom";
 
 const GlobalNav = () => {
-    return(
+    return (
         <nav>
             <ul>
                 <Link to="/">
@@ -13,7 +13,7 @@ const GlobalNav = () => {
                 </Link>
             </ul>
         </nav>
-    )
-}
+    );
+};
 
 export default GlobalNav;
