+++ resources/js/components/MakeMailTree.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import EachMail from "./EachMail";
+
+function MakeMailTree(props) {
+    const mailtp = props.mailtp;
+    const tp = mailtp.map((mail) => <EachMail key={mail} mail={mail} />);
+
+    const result = tp;
+    return result;
+}
+
+export default MakeMailTree;
