+++ resources/js/components/MakeTree.jsx
@@ -1,0 +1,23 @@
+import React from "react";
+import SplitPane from "react-split-pane";
+import Typography from "@material-ui/core/Typography";
+import MakeMailTree from "./MakeMailTree";
+
+function MakeTree(props) {
+    const mails = props.mails;
+    //const mailItems = mails.map((mail) => <EachMail mail={mail} />);
+    const mailTrees = <MakeMailTree mailtp={mails} />;
+
+    //return <EachMail mail={props.mailList} />;
+    const layout = (
+        <SplitPane split="vertical" minSize={200} defaultSize={500} maxSize={1000}>
+            {mailTrees}
+            <Typography variant="h5" component="h3">
+                This is a sheet of paper.
+            </Typography>
+        </SplitPane>
+    );
+    return layout;
+}
+
+export default MakeTree;
